Create and fill bidskit.translate module
